requirements.conda.txt
blas==2.121
blas-devel==3.9.0
brotli-python==1.1.0
bzip2==1.0.8
ca-certificates==2024.8.30
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cuda-cudart==11.7.99
cuda-cupti==11.7.101
cuda-libraries==11.7.1
cuda-nvrtc==11.7.99
cuda-nvtx==11.7.91
cuda-runtime==11.7.1
cuda-version==12.6
cycler==0.12.1
ffmpeg==4.3
freetype==2.12.1
gmp==6.3.0
gnutls==3.6.13
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
icu==73.2
idna==3.10
imageio==2.36.0
joblib==1.4.2
kiwisolver==1.4.7
lame==3.100
lazy-loader==0.4
lcms2==2.15
ld_impl_linux-64==2.43
lerc==4.0.0
libblas==3.9.0
libcblas==3.9.0
libcublas==11.10.3.66
libcufft==10.7.2.124
libcufile==1.11.1.6
libcurand==10.3.7.77
libcusolver==11.4.0.1
libcusparse==11.7.4.91
libdeflate==1.17
libffi==3.4.2
libgcc==14.1.0
libgcc-ng==14.1.0
libgfortran==14.1.0
libgfortran-ng==14.1.0
libgfortran5==14.1.0
libgomp==14.1.0
libhwloc==2.11.1
libiconv==1.17
liblapack==3.9.0
liblapacke==3.9.0
libnpp==11.7.4.75
libnsl==2.0.1
libnvjpeg==11.8.0.2
libpng==1.6.43
libsqlite==3.46.0
libstdcxx==14.1.0
libstdcxx-ng==14.1.0
libtiff==4.5.0
libuuid==2.38.1
libwebp-base==1.4.0
libxcb==1.13
libxcrypt==4.4.36
libxml2==2.12.7
libzlib==1.2.13
llvm-openmp==19.1.1
matplotlib==3.1.3
mkl==2024.0.0
mkl-devel==2024.0.0
mkl-include==2024.0.0
ncurses==6.5
nettle==3.6
networkx==3.2.1
numpy==1.26.4
opencv-python==3.4.18.65
openh264==2.1.1
openjpeg==2.5.0
openssl==3.3.2
packaging==24.1
pillow==9.4.0
pip==24.2
pthread-stubs==0.4
pyaml==24.9.0
pycparser==2.22
pyparsing==3.2.0
pysocks==1.7.1
python==3.9.19
python-dateutil==2.9.0.post0
python_abi==3.9
pytorch==1.13.1
pytorch-cuda==11.7
pytorch-mutex==1.0
pyyaml==6.0.2
readline==8.2
requests==2.32.3
scikit-image==0.24.0
scikit-learn==0.24.2
scipy==1.13.1
setuptools==75.1.0
six==1.16.0
sqlite==3.46.0
tbb==2021.13.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tk==8.6.13
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.12.2
tzdata==2024b
urllib3==2.2.3
wheel==0.44.0
xorg-libxau==1.0.11
xorg-libxdmcp==1.1.5
xz==5.2.6
zlib==1.2.13
zstandard==0.23.0
zstd==1.5.6
